refactored vampire specific code to fit into the atomate2 folder scheme added a ex_mat field to the ExchangeDocument
