+++ src/atomate2/common/jobs/exchange.py
@@ -14,14 +14,10 @@ from jobflow import job
 from pymatgen.analysis.magnetism.heisenberg import HeisenbergMapper
 
 from atomate2.common.schemas.exchange import ExchangeDocument
-from atomate2.vampire import VampireCaller
 
 if TYPE_CHECKING:
     from pymatgen.analysis.magnetism.heisenberg import HeisenbergModel
     from pymatgen.core.structure import Structure
-
-    from atomate2.vampire import VampireOutput
-
 
 logger = logging.getLogger(__name__)
 
@@ -57,35 +53,6 @@ def heisenberg_mapping(
     total_energies = [e * len(s) for s, e in zip(structures, energies, strict=True)]
     hmapper = HeisenbergMapper(structures, total_energies, **heisenberg_settings)
     return hmapper.get_heisenberg_model()
-
-
-@job(name="run vampire")
-def run_vampire(
-    heisenberg_model: HeisenbergModel,
-    mc_settings: dict | None = None,
-) -> VampireOutput:
-    """Run Vampire Monte-Carlo to estimate the critical temperature.
-
-    This wraps the (vendored) ``VampireCaller``, which shells out to the external
-    ``vampire-serial`` binary. A clear error is raised if the binary is not found
-    on PATH.
-
-    Parameters
-    ----------
-    heisenberg_model : HeisenbergModel
-        The fitted Heisenberg model from :func:`heisenberg_mapping`.
-    mc_settings : dict or None
-        Keyword arguments for VampireCaller, e.g. ``mc_box_size``,
-        ``equil_timesteps``, ``mc_timesteps``, ``avg``.
-
-    Returns
-    -------
-    VampireOutput
-        The Vampire Monte-Carlo result, exposing ``critical_temp``.
-    """
-    mc_settings = mc_settings or {}
-    vampire_caller = VampireCaller(hm=heisenberg_model, **mc_settings)
-    return vampire_caller.output
 
 
 @job(name="build exchange doc")
